requirements.txt
praw==7.8.1
groq==0.30.0
python-dotenv==1.1.1
streamlit==1.46.1
openai==1.3.7
requests==2.32.4
beautifulsoup4==4.12.2
lxml==4.9.3
pandas==2.3.1
numpy==2.2.6
pillow
plotly==6.2.0 